04_RAG_AND_AI_AGENTS/01_BASIC_RAG/STAGE_3_APP_lvl2/BE/requirements.txt
fastapi
google-generativeai
python-dotenv
uvicorn
pymilvus
marshmallow==3.20.1
google-genai